fix the bots
